+++ src/components/Comments/Comment.css
@@ -1,4 +1,4 @@
-.ProposalComment {
+.Comment {
   display: flex;
   flex-direction: row;
   justify-content: flex-start;
@@ -7,60 +7,60 @@
   margin-bottom: 24px;
 }
 
-.ProposalComment .ProposalComment__ProfileImage {
+.Comment__ProfileImage {
   margin-right: 16px;
 }
 
-.ProposalComment .ProposalComment__Author {
+.Comment__Author {
   display: flex;
   align-items: baseline;
   justify-content: flex-start;
   margin-bottom: 15px;
 }
 
-.ProposalComment .ProposalComment__Author a {
+.Comment__Author a {
   font-weight: var(--weight-semi-bold);
 }
 
-.ProposalComment__Cooked {
+.Comment__Cooked {
   width: auto;
   height: auto;
   max-width: 100%;
 }
 
-.ProposalComment__Cooked img {
+.Comment__Cooked img {
   max-width: 100%;
   height: auto;
 }
 
-.ProposalComment__Cooked__Img {
+.Comment__CookedImg {
   width: fit-content;
   height: fit-content;
   max-width: 100%;
   margin-right: 1rem;
 }
 
-.ProposalComment span {
+.Comment span {
   flex: 1 1;
   line-height: 20px;
   margin: 0 7px;
 }
 
-.ProposalComment div.title {
+.Comment div.title {
   display: flex;
   flex-direction: row;
   align-items: center;
 }
 
-.ProposalComments img.emoji {
+.Comments img.emoji {
   vertical-align: middle;
 }
 
-.ProposalComments img.avatar {
+.Comments img.avatar {
   margin-right: 10px;
 }
 
-.ProposalComments aside.quote {
+.Comments aside.quote {
   display: flex;
   flex-direction: column;
   border-left: 5px solid var(--black-600);
@@ -70,6 +70,6 @@
   margin-bottom: 1em;
 }
 
-.ProposalComment .ProposalComment__Content {
+.Comment .Comment__Content {
   width: 100%;
 }
