+++ src/components/Comments/Comment.tsx
@@ -3,16 +3,16 @@ import React from 'react'
 import DOMPurify from 'dompurify'
 import isEthereumAddress from 'validator/lib/isEthereumAddress'
 
-import { FORUM_URL } from '../../../constants'
-import useProfile from '../../../hooks/useProfile'
-import Time from '../../../utils/date/Time'
-import locations from '../../../utils/locations'
-import Avatar from '../../Common/Avatar'
-import Link from '../../Common/Typography/Link'
-import Text from '../../Common/Typography/Text'
-import ValidatedProfile from '../../Icon/ValidatedProfile'
+import { FORUM_URL } from '../../constants'
+import useProfile from '../../hooks/useProfile'
+import Time from '../../utils/date/Time'
+import locations from '../../utils/locations'
+import Avatar from '../Common/Avatar'
+import Link from '../Common/Typography/Link'
+import Text from '../Common/Typography/Text'
+import ValidatedProfile from '../Icon/ValidatedProfile'
 
-import './ProposalComment.css'
+import './Comment.css'
 
 function getUserProfileUrl(user: string, address?: string) {
   return address ? locations.profile({ address }) : `${FORUM_URL}/u/${user}`
@@ -26,11 +26,11 @@ type Props = {
   address?: string
 }
 
-export default function ProposalComment({ user, avatarUrl, createdAt, cooked, address }: Props) {
+export default function Comment({ user, avatarUrl, createdAt, cooked, address }: Props) {
   const createMarkup = (html: any) => {
     DOMPurify.addHook('afterSanitizeAttributes', function (node) {
       if (node.nodeName && node.nodeName === 'IMG' && node.getAttribute('alt') === 'image') {
-        node.className = 'ProposalComment__Cooked__Img'
+        node.className = 'Comment__CookedImg'
       }
 
       const hrefAttribute = node.getAttribute('href')
@@ -52,14 +52,14 @@ export default function ProposalComment({ user, avatarUrl, createdAt, cooked, ad
   const linkRel = address ? undefined : 'noopener noreferrer'
 
   return (
-    <div className="ProposalComment">
-      <div className="ProposalComment__ProfileImage">
+    <div className="Comment">
+      <div className="Comment__ProfileImage">
         <Link href={discourseUserUrl} target={linkTarget} rel={linkRel}>
           {address ? <Avatar address={address} size="medium" /> : <Avatar size="medium" src={avatarUrl} />}
         </Link>
       </div>
-      <div className="ProposalComment__Content">
-        <div className="ProposalComment__Author">
+      <div className="Comment__Content">
+        <div className="Comment__Author">
           <Link href={discourseUserUrl} target={linkTarget} rel={linkRel}>
             <Text weight="bold">
               {displayableAddress && !isEthereumAddress(displayableAddress) ? displayableAddress : user}
@@ -70,7 +70,7 @@ export default function ProposalComment({ user, avatarUrl, createdAt, cooked, ad
             <Text color="secondary">{Time.from(createdAt).fromNow()}</Text>
           </span>
         </div>
-        <div className="ProposalComment__Cooked" dangerouslySetInnerHTML={createMarkup(cooked)} />
+        <div className="Comment__Cooked" dangerouslySetInnerHTML={createMarkup(cooked)} />
       </div>
     </div>
   )
