+++ src/components/Comments/Comments.css
@@ -2,7 +2,7 @@
   margin-top: 36px;
 }
 
-.ProposalComment__Author svg {
+.Comment__Author svg {
   margin-left: 4px;
 }
