+++ src/components/Proposal/Comments/ProposalComments.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 
 import { ProposalAttributes } from '../../../entities/Proposal/types'
 import useProposalComments from '../../../hooks/useProposalComments'
-
-import Comments from './Comments'
+import Comments from '../../Comments/Comments'
 
 type ProposalComments = {
   proposal: ProposalAttributes | null
@@ -18,6 +17,7 @@ export default function ProposalComments({ proposal }: ProposalComments) {
       isLoading={isLoadingComments}
       topicId={proposal?.discourse_topic_id}
       topicSlug={proposal?.discourse_topic_slug}
+      topicType="proposal"
     />
   )
 }
